refactor(interpreter): split variable and assignment statements
