+++ src/interpreter/src/statements/mod.rs
@@ -6,6 +6,27 @@ use paste::paste;
 #[cfg(feature = "variable_define")]
 use crate::stdlib::define::*;
 
+mod op_assign;
+mod variable_assign;
+mod variable_define;
+
+#[cfg(feature = "math")]
+pub use op_assign::op_assign;
+#[cfg(feature = "math_add_assign")]
+pub use op_assign::add_assign;
+#[cfg(feature = "math_sub_assign")]
+pub use op_assign::sub_assign;
+#[cfg(feature = "math_div_assign")]
+pub use op_assign::mul_assign;
+#[cfg(feature = "math_mul_assign")]
+pub use op_assign::div_assign;
+#[cfg(feature = "variable_assign")]
+pub use variable_assign::variable_assign;
+#[cfg(all(feature = "subscript", feature = "assign"))]
+pub use variable_assign::subscript_ref;
+#[cfg(feature = "variable_define")]
+pub use variable_define::variable_define;
+
 // Statements
 // ----------------------------------------------------------------------------
 
@@ -110,141 +131,6 @@ pub fn tuple_destructure(tpl_dstrct: &TupleDestructure, p: &InterpreterExecution
     }
   }
   Ok(source)
-}
-
-fn assignment_registration_operand(value: &Value) -> Value {
-  match value {
-    Value::MutableReference(reference) => {
-      let inner = reference.borrow();
-      assignment_registration_operand(&inner)
-    }
-    _ => value.clone(),
-  }
-}
-
-#[cfg(feature = "math")]
-pub fn op_assign(op_assgn: &OpAssign, env: Option<&Environment>, p: &InterpreterExecution<'_>) -> MResult<Value> {
-  let mut source = expression(&op_assgn.expression, env, p)?;
-  let slc = &op_assgn.target;
-  if slc.context.is_some() {
-    return Err(MechError::new(AddressedAssignmentUnsupported, None)
-      .with_compiler_loc()
-      .with_tokens(slc.tokens()));
-  }
-  let id = slc.name.hash();
-  let sink = {
-    let mut state_brrw = p.state.borrow_mut();
-    match state_brrw.get_mutable_symbol(id) {
-      Some(val) => val.borrow().clone(),
-      None => {
-        match state_brrw.contains_symbol(id) {
-          true => return Err(MechError::new(
-            NotMutableError { id },
-            Some("(!)> Mutable variables are defined with the `~` operator. *e.g.*: {{~x := 123}}".to_string()),
-          ).with_compiler_loc().with_tokens(slc.name.tokens())),
-          false => return Err(MechError::new(
-            UndefinedVariableError { id, name: slc.name.to_string() },
-            Some("(!)> Variables are defined with the `:=` operator. *e.g.*: {{x := 123}}".to_string()),
-          ).with_compiler_loc().with_tokens(slc.name.tokens())),
-        }
-      }
-    }
-  };
-  match &slc.subscript {
-    Some(sbscrpt) => {
-      // todo: this only works for the first subscript, it needs to work for multiple subscripts
-      for s in sbscrpt {
-        let fxn = match op_assgn.op {
-          #[cfg(feature = "math_add_assign")]
-          OpAssignOp::Add => add_assign(&s, &sink, &source, env, p)?,
-          #[cfg(feature = "math_sub_assign")]
-          OpAssignOp::Sub => sub_assign(&s, &sink, &source, env, p)?,
-          #[cfg(feature = "math_div_assign")]
-          OpAssignOp::Div => div_assign(&s, &sink, &source, env, p)?,
-          #[cfg(feature = "math_mul_assign")]
-          OpAssignOp::Mul => mul_assign(&s, &sink, &source, env, p)?,
-          _ => todo!(),
-        };
-        return Ok(fxn);
-      }
-    }
-    None => {
-      let plan = p.plan();
-      let registration_source = assignment_registration_operand(&source);
-      let compile_arguments = vec![sink, source];
-      let registration_arguments = vec![registration_source];
-      return match op_assgn.op {
-        #[cfg(feature = "math_add_assign")]
-        OpAssignOp::Add => execute_initialized_indexed_compiler_with_registration_arguments(p, &plan, &AddAssignValue{}, compile_arguments, registration_arguments),
-        #[cfg(feature = "math_sub_assign")]
-        OpAssignOp::Sub => execute_initialized_indexed_compiler_with_registration_arguments(p, &plan, &SubAssignValue{}, compile_arguments, registration_arguments),
-        #[cfg(feature = "math_div_assign")]
-        OpAssignOp::Div => execute_initialized_indexed_compiler_with_registration_arguments(p, &plan, &DivAssignValue{}, compile_arguments, registration_arguments),
-        #[cfg(feature = "math_mul_assign")]
-        OpAssignOp::Mul => execute_initialized_indexed_compiler_with_registration_arguments(p, &plan, &MulAssignValue{}, compile_arguments, registration_arguments),
-        _ => todo!(),
-      };
-    }
-  }
-  unreachable!(); // subscript should have thrown an error if we can't access an element
-}
-
-#[cfg(feature = "variable_assign")]
-pub fn variable_assign(var_assgn: &VariableAssign, env: Option<&Environment>, p: &InterpreterExecution<'_>) -> MResult<Value> {
-  let mut source = expression(&var_assgn.expression, env, p)?;
-  let slc = &var_assgn.target;
-  if slc.context.is_some() {
-    return Err(MechError::new(AddressedAssignmentUnsupported, None)
-      .with_compiler_loc()
-      .with_tokens(slc.tokens()));
-  }
-  let id = slc.name.hash();
-  let sink = {
-    let symbols = p.symbols();
-    let symbols_brrw = symbols.borrow();
-    match symbols_brrw.get_mutable(id) {
-      Some(val) => val.borrow().clone(),
-      None => {
-        if !symbols_brrw.contains(id) {
-          return Err(MechError::new(
-            UndefinedVariableError { id, name: slc.name.to_string() },
-            Some("(!)> Variables are defined with the `:=` operator. *e.g.*: {{x := 123}}".to_string()),
-          ).with_compiler_loc().with_tokens(slc.name.tokens()));
-        } else {
-          return Err(MechError::new(
-            NotMutableError { id },
-            Some("(!)> Mutable variables are defined with the `~` operator. *e.g.*: {{~x := 123}}".to_string()),
-          ).with_compiler_loc().with_tokens(slc.name.tokens()));
-        }
-      }
-    }
-  };
-  match &slc.subscript {
-    Some(sbscrpt) => {
-      #[cfg(feature = "subscript")]
-      for s in sbscrpt {
-        let s_result = subscript_ref(&s, &sink, &source, env, p)?;
-        return Ok(s_result);
-      }
-    }
-    #[cfg(feature = "assign")]
-    None => {
-      let plan = p.plan();
-      let registration_source = assignment_registration_operand(&source);
-      return execute_initialized_indexed_compiler_with_registration_arguments(
-        p,
-        &plan,
-        &AssignValue{},
-        vec![sink, source],
-        vec![registration_source],
-      );
-    }
-    _ => return Err(MechError::new(
-      FeatureNotEnabledError,
-      None
-    ).with_compiler_loc().with_tokens(var_assgn.target.tokens())),
-  }
-  unreachable!(); // subscript should have thrown an error if we can't access an element
 }
 
 #[cfg(feature = "enum")]
@@ -518,135 +404,6 @@ fn value_matches_enum_variant(value: &Value, enum_id: u64, state: &ProgramState)
   }
 }
 
-#[cfg(feature = "variable_define")]
-pub fn variable_define(var_def: &VariableDefine, p: &InterpreterExecution<'_>) -> MResult<Value> {
-  if var_def.var.context.is_some() {
-    return Err(MechError::new(AddressedAssignmentUnsupported, None)
-      .with_compiler_loc()
-      .with_tokens(var_def.var.tokens()));
-  }
-  let var_id = var_def.var.name.hash();
-  let var_name = var_def.var.name.to_string();
-  {
-    let symbols = p.symbols();
-    if symbols.borrow().contains(var_id) {
-      return Err(MechError::new(
-        VariableAlreadyDefinedError { id: var_id },
-        None
-      ).with_compiler_loc().with_tokens(var_def.var.name.tokens()));
-    }
-  }
-  let plan = p.plan();
-  #[cfg(feature = "subscript_formula")]
-  reset_current_string_access_expression_live(p);
-  let mut result = expression(&var_def.expression, None, p)?;
-  #[cfg(feature = "subscript_formula")]
-  let string_access_result_is_live = take_current_string_access_expression_live(p);
-  #[cfg(all(feature = "kind_annotation", feature = "convert"))]
-  if let Some(knd_anntn) =  &var_def.var.kind {
-    let knd = kind_annotation(&knd_anntn.kind,p)?;
-    let mut state_brrw = &mut p.state.borrow_mut();
-    let target_knd = knd.to_value_kind(&mut state_brrw.kinds)?;
-    // Do kind checking
-    match (&result, &target_knd) {
-      // Atom is a variant of an enum
-      #[cfg(all(feature = "atom", feature = "enum"))]
-      (Value::Atom(atom_variant), ValueKind::Enum(enum_id, target_enum_variant_name)) => {
-        let atom_name = atom_variant.borrow().name();
-        if !value_matches_enum_variant(&result, *enum_id, &*state_brrw) {
-          return Err(MechError::new(
-            UnableToConvertAtomToEnumVariantError { atom_name: atom_name.clone(), target_enum_variant_name: target_enum_variant_name.clone() },
-            None
-          ).with_compiler_loc().with_tokens(var_def.expression.tokens()));
-        }
-      }
-      #[cfg(all(feature = "tuple", feature = "atom", feature = "enum"))]
-      (Value::Tuple(tuple_val), ValueKind::Enum(enum_id, target_enum_variant_name)) => {
-        let atom_name = format!("{:?}", tuple_val);
-        if !value_matches_enum_variant(&result, *enum_id, &*state_brrw) {
-          return Err(MechError::new(
-            UnableToConvertAtomToEnumVariantError { atom_name, target_enum_variant_name: target_enum_variant_name.clone() },
-            None
-          ).with_compiler_loc().with_tokens(var_def.expression.tokens()));
-        }
-      }
-      // Atoms can't convert into anything else.
-      #[cfg(feature = "atom")]
-      (Value::Atom(given_variant_id), target_kind) => {
-        return Err(MechError::new(
-          UnableToConvertAtomError { atom_id: given_variant_id.borrow().0.0},
-          None
-        ).with_compiler_loc().with_tokens(var_def.expression.tokens()));
-      }
-      #[cfg(feature = "record")]
-      (Value::Record(rec), ref target_kind @ ValueKind::Record(target_rec_knd)) => {
-        let rec_brrw = rec.borrow();
-        let rec_knd = rec_brrw.kind();
-        if &rec_knd != *target_kind {
-          return Err(MechError::new(
-            UnableToConvertRecordError { source_record_kind: rec_knd.clone(), target_record_kind: (*target_kind).clone() },
-            None
-          ).with_compiler_loc().with_tokens(var_def.expression.tokens()));
-        }
-      }
-      #[cfg(feature = "matrix")]
-      (Value::MutableReference(v), ValueKind::Matrix(target_matrix_knd,_)) => {
-        let value = v.borrow().clone();
-        if value.is_matrix() {
-          result = execute_initialized_indexed_compiler(p, &plan, &ConvertMatToMat{}, vec![result.clone(), Value::Kind(target_knd.clone())])?;
-        } else {
-          let value_kind = value.kind();
-          if value_kind.deref_kind() != target_matrix_knd.as_ref().clone() && value_kind != *target_matrix_knd.clone() {
-            result = execute_initialized_indexed_compiler(p, &plan, &ConvertKind{}, vec![result.clone(), Value::Kind(target_matrix_knd.as_ref().clone())])?;
-          };
-          result = execute_initialized_indexed_compiler(p, &plan, &ConvertScalarToMat{}, vec![result.clone(), Value::Kind(target_knd.clone())])?;
-        }
-      }
-      #[cfg(feature = "matrix")]
-      (value, ValueKind::Matrix(target_matrix_knd,_)) => {
-        if value.is_matrix() {
-          result = execute_initialized_indexed_compiler(p, &plan, &ConvertMatToMat{}, vec![result.clone(), Value::Kind(target_knd.clone())])?;
-        } else {
-          let value_kind = value.kind();
-          if value_kind.deref_kind() != target_matrix_knd.as_ref().clone() && value_kind != *target_matrix_knd.clone() {
-            result = execute_initialized_indexed_compiler(p, &plan, &ConvertKind{}, vec![result.clone(), Value::Kind(target_matrix_knd.as_ref().clone())])?;
-          };
-          result = execute_initialized_indexed_compiler(p, &plan, &ConvertScalarToMat{}, vec![result.clone(), Value::Kind(target_knd.clone())])?;
-        }
-      }
-      // Kind isn't checked
-      x => {
-        result = execute_initialized_indexed_compiler(p, &plan, &ConvertKind{}, vec![result.clone(), Value::Kind(target_knd)])?;
-      },
-    };
-    let detached_result = detach_variable_value(&result);
-    #[cfg(feature = "subscript_formula")]
-    if string_access_result_is_live {
-      mark_string_access_value_live(p, &detached_result);
-    }
-    // Save symbol to interpreter
-    let val_ref = state_brrw.save_symbol(var_id, var_name.clone(), detached_result.clone(), var_def.mutable);
-    // Add variable define step to plan
-    let var_define_arguments = vec![detached_result.clone(), Value::String(Ref::new(var_name.clone())), Value::Bool(Ref::new(var_def.mutable))];
-    let var_def_fxn = VarDefine{}.compile(&var_define_arguments)?;
-    plan.register_function(var_def_fxn, &[])?;
-    return Ok(detached_result);
-  }
-  let mut state_brrw = p.state.borrow_mut();
-  let detached_result = detach_variable_value(&result);
-  #[cfg(feature = "subscript_formula")]
-  if string_access_result_is_live {
-    mark_string_access_value_live(p, &detached_result);
-  }
-  // Save symbol to interpreter
-  let val_ref = state_brrw.save_symbol(var_id,var_name.clone(),detached_result.clone(),var_def.mutable);
-  // Add variable define step to plan
-  let var_define_arguments = vec![detached_result.clone(), Value::String(Ref::new(var_name.clone())), Value::Bool(Ref::new(var_def.mutable))];
-  let var_def_fxn = VarDefine{}.compile(&var_define_arguments)?;
-  plan.register_function(var_def_fxn, &[])?;
-  return Ok(detached_result);
-}
-
 #[cfg(feature = "state_machines")]
 pub fn fsm_declare(fsm_decl: &FsmDeclare, env: Option<&Environment>, p: &InterpreterExecution<'_>) -> MResult<Value> {
   let result = crate::state_machines::execute_fsm_pipe(&fsm_decl.pipe, env, p)?;
@@ -668,321 +425,6 @@ fn detach_variable_value(value: &Value) -> Value {
     _ => value.clone(),
   }
 }
-
-macro_rules! op_assign {
-  ($fxn_name:ident, $op:tt) => {
-    paste!{
-      pub fn $fxn_name(sbscrpt: &Subscript, sink: &Value, source: &Value, env: Option<&Environment>, p: &InterpreterExecution<'_>) -> MResult<Value> {
-        let plan = p.plan();
-        match sbscrpt {
-          Subscript::Dot(x) => {
-            todo!()
-          },
-          Subscript::DotInt(x) => {
-            todo!()
-          },
-          Subscript::Swizzle(x) => {
-            todo!()
-          },
-          Subscript::Bracket(subs) => {
-            let mut fxn_input = vec![sink.clone()];
-            match &subs[..] {
-              [Subscript::Formula(ix)] => {
-                fxn_input.push(source.clone());
-                let ixes = subscript_formula_ix(&subs[0], env, p)?;
-                let shape = ixes.shape();
-                fxn_input.push(ixes);
-                match shape[..] {
-                  [1,1] => { plan.borrow_mut().push(MatrixAssignScalar{}.compile(&fxn_input)?); }
-                  [1,n] => { plan.borrow_mut().push([<$op AssignRange>]{}.compile(&fxn_input)?); }
-                  [n,1] => { plan.borrow_mut().push([<$op AssignRange>]{}.compile(&fxn_input)?); }
-                  _ => todo!(),
-                }
-              },
-              [Subscript::Formula(ix1),Subscript::All] => {
-                fxn_input.push(source.clone());
-                let ix = subscript_formula_ix(&subs[0], env, p)?;
-                let shape = ix.shape();
-                fxn_input.push(ix);
-                fxn_input.push(Value::IndexAll);
-                match shape[..] {
-                  [1,1] => { plan.borrow_mut().push(MatrixAssignScalarAll{}.compile(&fxn_input)?); }
-                  [1,n] => { plan.borrow_mut().push([<$op AssignRangeAll>]{}.compile(&fxn_input)?); }
-                  [n,1] => { plan.borrow_mut().push([<$op AssignRangeAll>]{}.compile(&fxn_input)?); }
-                  _ => todo!(),
-                }
-              },
-              [Subscript::Range(ix)] => {
-                fxn_input.push(source.clone());
-                let ixes = subscript_range(&subs[0], env, p)?;
-                fxn_input.push(ixes);
-                plan.borrow_mut().push([<$op AssignRange>]{}.compile(&fxn_input)?);
-              },
-              [Subscript::Range(ix), Subscript::All] => {
-                fxn_input.push(source.clone());
-                let ixes = subscript_range(&subs[0], env, p)?;
-                fxn_input.push(ixes);
-                fxn_input.push(Value::IndexAll);
-                plan.borrow_mut().push([<$op AssignRangeAll>]{}.compile(&fxn_input)?);
-              },
-              x => todo!("{:?}", x),
-            };
-            let plan_brrw = plan.borrow();
-            let mut new_fxn = &plan_brrw.last().unwrap();
-            new_fxn.solve();
-            let res = new_fxn.out();
-            return Ok(res);
-          },
-          Subscript::Brace(x) => todo!(),
-          x => todo!("{:?}", x),
-        }
-      }
-    }
-  };
-}
-
-#[cfg(feature = "math_add_assign")]
-op_assign!(add_assign, Add);
-#[cfg(feature = "math_sub_assign")]
-op_assign!(sub_assign, Sub);
-#[cfg(feature = "math_div_assign")]
-op_assign!(mul_assign, Mul);
-#[cfg(feature = "math_mul_assign")]
-op_assign!(div_assign, Div);
-//#[cfg(feature = "math_pow")]
-//op_assign!(pow_assign, Pow);
-
-#[cfg(all(feature = "subscript", feature = "assign"))]
-pub fn subscript_ref(sbscrpt: &Subscript, sink: &Value, source: &Value, env: Option<&Environment>, p: &InterpreterExecution<'_>) -> MResult<Value> {
-  let plan = p.plan();
-  let symbols = p.symbols();
-  let functions = p.functions();
-  match sbscrpt {
-    Subscript::Dot(x) => {
-      let key = x.hash();
-      let fxn_input: Vec<Value> = vec![sink.clone(), source.clone(), Value::Id(key)];
-      let new_fxn = AssignColumn{}.compile(&fxn_input)?;
-      new_fxn.solve();
-      let res = new_fxn.out();
-      plan.borrow_mut().push(new_fxn);
-      return Ok(res);
-    },
-    #[cfg(feature = "tuple")]
-    Subscript::DotInt(x) => {
-      let ix = real(x, p)?.as_index()?;
-      let mut fxn_input: Vec<Value> = vec![sink.clone(), source.clone(), ix.clone()];
-      let new_fxn = TupleAssignScalar{}.compile(&fxn_input)?;
-      new_fxn.solve();
-      let res = new_fxn.out();
-      plan.borrow_mut().push(new_fxn);
-      return Ok(res);
-    },
-    Subscript::Swizzle(x) => {
-      unreachable!()
-    },
-    Subscript::Bracket(subs) => {
-      let mut fxn_input = vec![sink.clone()];
-      match &subs[..] {
-        #[cfg(feature = "subscript_formula")]
-        [Subscript::Formula(ix)] => {
-          fxn_input.push(source.clone());
-          let ixes = subscript_formula_ix(&subs[0], env, p)?;
-          let shape = ixes.shape();
-          fxn_input.push(ixes);
-          match shape[..] {
-            #[cfg(feature = "matrix")]
-            [1,1] => { plan.borrow_mut().push(MatrixAssignScalar{}.compile(&fxn_input)?); }
-            #[cfg(all(feature = "matrix", feature = "subscript_range", feature = "assign"))]
-            [1,n] => { plan.borrow_mut().push(MatrixAssignRange{}.compile(&fxn_input)?); }
-            #[cfg(all(feature = "matrix", feature = "subscript_range", feature = "assign"))]
-            [n,1] => { plan.borrow_mut().push(MatrixAssignRange{}.compile(&fxn_input)?); }
-            _ => todo!(),
-          }
-        },
-        #[cfg(all(feature = "matrix", feature = "subscript_range"))]
-        [Subscript::Range(ix)] => {
-          fxn_input.push(source.clone());
-          let ixes = subscript_range(&subs[0], env, p)?;
-          fxn_input.push(ixes);
-          plan.borrow_mut().push(MatrixAssignRange{}.compile(&fxn_input)?);
-        },
-        #[cfg(all(feature = "matrix", feature = "subscript_range"))]
-        [Subscript::All] => {
-          fxn_input.push(source.clone());
-          fxn_input.push(Value::IndexAll);
-          plan.borrow_mut().push(MatrixAssignAll{}.compile(&fxn_input)?);
-        },
-        [Subscript::All,Subscript::All] => todo!(),
-        #[cfg(feature = "subscript_formula")]
-        [Subscript::Formula(ix1),Subscript::Formula(ix2)] => {
-          fxn_input.push(source.clone());
-          let result1 = subscript_formula_ix(&subs[0], env, p)?;
-          let result2 = subscript_formula_ix(&subs[1], env, p)?;
-          let shape1 = result1.shape();
-          let shape2 = result2.shape();
-          fxn_input.push(result1);
-          fxn_input.push(result2);
-          match ((shape1[0],shape1[1]),(shape2[0],shape2[1])) {
-            #[cfg(feature = "matrix")]
-            ((1,1),(1,1)) => { plan.borrow_mut().push(MatrixAssignScalarScalar{}.compile(&fxn_input)?); }
-            #[cfg(all(feature = "matrix", feature = "subscript_range"))]
-            ((1,1),(m,1)) => { plan.borrow_mut().push(MatrixAssignScalarRange{}.compile(&fxn_input)?); }
-            #[cfg(all(feature = "matrix", feature = "subscript_range"))]
-            ((n,1),(1,1)) => { plan.borrow_mut().push(MatrixAssignRangeScalar{}.compile(&fxn_input)?); }
-            #[cfg(all(feature = "matrix", feature = "subscript_range"))]
-            ((n,1),(m,1)) => { plan.borrow_mut().push(MatrixAssignRangeRange{}.compile(&fxn_input)?); }
-            _ => unreachable!(),
-          }
-        },
-        #[cfg(all(feature = "matrix", feature = "subscript_range"))]
-        [Subscript::Range(ix1),Subscript::Range(ix2)] => {
-          fxn_input.push(source.clone());
-          let result = subscript_range(&subs[0], env, p)?;
-          fxn_input.push(result);
-          let result = subscript_range(&subs[1], env, p)?;
-          fxn_input.push(result);
-          plan.borrow_mut().push(MatrixAssignRangeRange{}.compile(&fxn_input)?);
-        },
-        #[cfg(all(feature = "matrix", feature = "subscript_formula"))]
-        [Subscript::All,Subscript::Formula(ix2)] => {
-          fxn_input.push(source.clone());
-          fxn_input.push(Value::IndexAll);
-          let ix = subscript_formula_ix(&subs[1], env, p)?;
-          let shape = ix.shape();
-          fxn_input.push(ix);
-          match shape[..] {
-            #[cfg(feature = "matrix")]
-            [1,1] => { plan.borrow_mut().push(MatrixAssignAllScalar{}.compile(&fxn_input)?); }
-            #[cfg(feature = "matrix")]
-            [1,n] => { plan.borrow_mut().push(MatrixAssignAllRange{}.compile(&fxn_input)?); }
-            #[cfg(feature = "matrix")]
-            [n,1] => { plan.borrow_mut().push(MatrixAssignAllRange{}.compile(&fxn_input)?); }
-            _ => todo!(),
-          }
-        }
-        #[cfg(feature = "subscript_formula")]
-        [Subscript::Formula(ix1),Subscript::All] => {
-          fxn_input.push(source.clone());
-          let ix = subscript_formula_ix(&subs[0], env, p)?;
-          let shape = ix.shape();
-          fxn_input.push(ix);
-          fxn_input.push(Value::IndexAll);
-          match shape[..] {
-            #[cfg(feature = "matrix")]
-            [1,1] => { plan.borrow_mut().push(MatrixAssignScalarAll{}.compile(&fxn_input)?); }
-            #[cfg(all(feature = "matrix", feature = "subscript_range"))]
-            [1,n] => { plan.borrow_mut().push(MatrixAssignRangeAll{}.compile(&fxn_input)?); }
-            #[cfg(all(feature = "matrix", feature = "subscript_range"))]
-            [n,1] => { plan.borrow_mut().push(MatrixAssignRangeAll{}.compile(&fxn_input)?); }
-            _ => todo!(),
-          }
-        },
-        #[cfg(all(feature = "subscript_formula", feature = "subscript_range"))]
-        [Subscript::Range(ix1),Subscript::Formula(ix2)] => {
-          fxn_input.push(source.clone());
-          let result = subscript_range(&subs[0], env, p)?;
-          fxn_input.push(result);
-          let result = subscript_formula_ix(&subs[1], env, p)?;
-          let shape = result.shape();
-          fxn_input.push(result);
-          match &shape[..] {
-            #[cfg(feature = "matrix")]
-            [1,1] => { plan.borrow_mut().push(MatrixAssignRangeScalar{}.compile(&fxn_input)?); }
-            #[cfg(feature = "matrix")]
-            [1,n] => { plan.borrow_mut().push(MatrixAssignRangeRange{}.compile(&fxn_input)?); }
-            #[cfg(feature = "matrix")]
-            [n,1] => { plan.borrow_mut().push(MatrixAssignRangeRange{}.compile(&fxn_input)?); }
-            _ => todo!(),
-          }
-        },
-        #[cfg(all(feature = "subscript_formula", feature = "subscript_range"))]
-        [Subscript::Formula(ix1),Subscript::Range(ix2)] => {
-          fxn_input.push(source.clone());
-          let result = subscript_formula_ix(&subs[0], env, p)?;
-          let shape = result.shape();
-          fxn_input.push(result);
-          let result = subscript_range(&subs[1], env, p)?;
-          fxn_input.push(result);
-          match &shape[..] {
-            #[cfg(feature = "matrix")]
-            [1,1] => { plan.borrow_mut().push(MatrixAssignScalarRange{}.compile(&fxn_input)?); }
-            #[cfg(feature = "matrix")]
-            [1,n] => { plan.borrow_mut().push(MatrixAssignRangeRange{}.compile(&fxn_input)?); }
-            #[cfg(feature = "matrix")]
-            [n,1] => { plan.borrow_mut().push(MatrixAssignRangeRange{}.compile(&fxn_input)?); }
-            _ => todo!(),
-          }
-        },
-        #[cfg(all(feature = "matrix", feature = "subscript_range"))]
-        [Subscript::All,Subscript::Range(ix2)] => {
-          fxn_input.push(source.clone());
-          fxn_input.push(Value::IndexAll);
-          let result = subscript_range(&subs[1], env, p)?;
-          fxn_input.push(result);
-          plan.borrow_mut().push(MatrixAssignAllRange{}.compile(&fxn_input)?);
-        },
-        #[cfg(all(feature = "matrix", feature = "subscript_range"))]
-        [Subscript::Range(ix1),Subscript::All] => {
-          fxn_input.push(source.clone());
-          let result = subscript_range(&subs[0], env, p)?;
-          fxn_input.push(result);
-          fxn_input.push(Value::IndexAll);
-          plan.borrow_mut().push(MatrixAssignRangeAll{}.compile(&fxn_input)?);
-        },
-        _ => unreachable!(),
-      };
-      let plan_brrw = plan.borrow();
-      let mut new_fxn = &plan_brrw.last().unwrap();
-      new_fxn.solve();
-      let res = new_fxn.out();
-      return Ok(res);
-    },
-    Subscript::Brace(subs) => {
-      let mut fxn_input = vec![sink.clone()];
-      match &subs[..] {
-        #[cfg(feature = "subscript_formula")]
-        [Subscript::Formula(ix)] => {
-          fxn_input.push(source.clone());
-          let ixes = subscript_formula(&subs[0], env, p)?;
-          let shape = ixes.shape();
-          fxn_input.push(ixes);
-          match shape[..] {
-            #[cfg(feature = "map")]
-            [1,1] => { plan.borrow_mut().push(MapAssignScalar{}.compile(&fxn_input)?); }
-            //#[cfg(all(feature = "matrix", feature = "subscript_range", feature = "assign"))]
-            //[1,n] => { plan.borrow_mut().push(MatrixAssignRange{}.compile(&fxn_input)?); }
-            //#[cfg(all(feature = "matrix", feature = "subscript_range", feature = "assign"))]
-            //[n,1] => { plan.borrow_mut().push(MatrixAssignRange{}.compile(&fxn_input)?); }
-            _ => todo!(),
-          }
-        },
-        #[cfg(all(feature = "matrix", feature = "subscript_range"))]
-        [Subscript::Range(ix)] => {
-          todo!();
-          //fxn_input.push(source.clone());
-          //let ixes = subscript_range(&subs[0], env, p)?;
-          //fxn_input.push(ixes);
-          //plan.borrow_mut().push(MatrixAssignRange{}.compile(&fxn_input)?);
-        },
-        #[cfg(all(feature = "matrix", feature = "subscript_range"))]
-        [Subscript::All] => {
-          todo!();
-          //fxn_input.push(source.clone());
-          //fxn_input.push(Value::IndexAll);
-          //plan.borrow_mut().push(MatrixAssignAll{}.compile(&fxn_input)?);
-        },
-        _ => unreachable!(),
-      };
-      let plan_brrw = plan.borrow();
-      let mut new_fxn = &plan_brrw.last().unwrap();
-      new_fxn.solve();
-      let res = new_fxn.out();
-      return Ok(res);
-    }
-    _ => unreachable!(),
-  }
-}
-
 
 #[derive(Debug, Clone)]
 pub struct AddressedAssignmentUnsupported;
